backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
anthropic
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client
pydantic